fix delete api calls

+++ deleteFromDB.mjs
@@ -1,6 +1,9 @@
+import { isNumber } from "./dataValidation.mjs"
 export default async (db,table,id)=>{
-    return db(table)
+    if(isNumber(id)){
+        return db(table)
     .where('id',id)
     .delete()
     .returning('id')
+    } 
 }
